store the name and "enabled" flag with each filter
It is often convienient to enable or disable an ettercap filter during
runtime; to achieve this, this patch adds a flag to the list data
structure, indicating whether packets should be passed to the filter
script. The filename is also stored inside the structure, allowing the
interface to pass it on to the user.