fixed comparison bug in ec_decode
the param pointer points to a struct iface_env object, thus the
comparisons against the GBL_OPTIONS->iface and GBL_OPTIONS->iface_bridge
strings are invalid. really we want to compare the global interface
names against ((struct iface_env *) param)->name, which is the name of
the interface from which libpcap got the current packet.

as far as i can tell, this bug was causing bridge mode to be entirely
broken, as the functions in ec_sniff_bridge.c rely on knowing which
interface from which the packet came. after playing around with it for
a bit, it looks like bridge mode now functions correctly.