Fix detection of system libnet.
The latest version of libnet from https://github.com/sam-github/libnet
has `#define LIBNET_VERSION  "1.2-rc2"`. This causes the current version
to not detect the presence of the system libnet because the regex used
to extract the version number only includes numbers and periods.

This fixes the detection of 1.2-rc* versions. No problems exist when
linking to the newer libnet library to the best of my knowledge.