Added a cmake linter
cmake-lint is a Python tool that checks and reports coding/style issues in CMake related build scripts.

Installation:
https://github.com/richq/cmake-lint
The recommended version is 1.4, which, at the time of this writing, was not yet published on PyPI.

From now on, ettercap uses cmake-lint (wrapped a shell script) to enforce a consistent, portable and up-to-date cmake coding practices. This means that all cmake files *contributed* to the project will undergo a series of style/portability checks before the travis build configuration phase can begin.

If this script finds any issues, it will exit with a non-zero status and a detailed error message. Else it won't output anything and exits with status code 0, continuing the build process.

The following rules are currently check for:

- Lines should be <= 120 characters long
Should be 80 lines but CMake versions prior to 3.0 do not support continuation with '\'.

- Expression repeated inside endif; use only endif()
...or only else(), endmacro(), endforeach(), endfunction() etc

- Do not use mixed case commands
It's all lowercase now

- Line ends in whitespace

- Extra spaces between 'some_function' and its ()

- Use (increments of) 2 spaces for indentation

- Use spaces for indentation not tabs

- Find modules should use uppercase names
So FindFOO.cmake not Findfoo.cmake

- Find Modules should include FindPackageHandleStandardArgs
Very useful. Also:

- Find Modules should use FIND_PACKAGE_HANDLE_STANDARD_ARGS

- Added script to report style/format issues in CMake files

- Updated .travis.yml to enforce said these rules
The very first command run by travis reports format/style issues and exits appropriately.
This test will only happen once per commit and in first matrix only. Not in every single Matrix.