Added solution for missing GTK2_GDKCONFIG_INCLUDE_DIR GTK2_GLIBCONFIG_INCLUDE_DIR
...on Windows.

GTK2 support on windows is just as abundant as the FindGTK2.cmake module.
It would take sending a patch upstream or hosting a local copy of FindGTK2.cmake in
cmake/Modules/ to fix it.
Neither of these steps is worth it. We'll look for GTK2 via pkg-config on windows from now.
All GTK2 bundles I know of ship pkg-config. So anyone on windows who wants GTK2, just needs to add
the GTK2 bundle's "bin" directory to their path. Easy peasy.