Temporary fixes for locating glibconfig.h (GTK3)
FindGTK3.cmake is terrible. 590 lines of confusion. I can already tell that I am
Going to hate that file. We're the only one's maintaining it and it didn't make it
to cmake trunk...

In addition, glibconfig.h is for some reason located under $PREFIX/lib/glib-2.0/include.
I failed to see why. FindGTK3.cmake is unable to locate it and I'm unable to teach
find_path() to look in "lib" directory for header files without using hard-coded pathnames.
Will need to come back to this. Added to TODO.CMake