Ship our own FindCurses.cmake and don't build the curses interface by default on windows
Turns out that the FindCurses.cmake module in cmake trunk has a bug where it only looks for libform,
even when you request wide-character support. It should look for libformw when CURSES_NEED_WIDE
is true.
That bug is not new. I saw that more than a year ago. Anyway until this is addressed,
we'll host our fixed copy of FindCurses.cmake.

Although these set of commits focus on enhancing cmake, I can already tell that the ncurses UI support
on Windows is promising yet a total mess, and *NOT* my priority.

To summarize:
As stated in include\missing\ncurses_mingw.h:
"OS_WINDOWS doesn't really use the full ncurses yet, but a small layer on-top of PD-curses."

So right now, we have a mix of a old PDcurses library, a local ncurses_mingw.h (which is never included)
some more local defines and functions that are now either part of the PDcurses library (vwprintw), have been
declared deprecated by it (PDC_check_bios_key) and cmake, eagerly looking for ncurses (and ncurses only).

Great.

The bad news is that PDcurses is much more limited then ncurses. The good news is that ncurses support on windows
came a long way. On windows, standard ncurses may very well work. So there may not be any need for PDcurses and any
of these "hacks".
But again, there's much more important things to worry about.
The mcurses interface will be off by default on windows

https://www.gnu.org/software/ncurses/#h4-port-mingw
http://invisible-island.net/ncurses/ncurses.html#download_mingw