Check for and set proper INSTALL_SYSCONFDIR
If we're targeting some flavor of UN*X or Cygwin (a unix-like environment on
windows with a POSIX-complaint shell and other GNU tools) set INSTALL_SYSCONFDIR
to "/etc", else set it to ${INSTALL_PREFIX}/etc.
That's mostly because there the root directory on those other OSs is not "/" and
it most certainly doesn't hold an "etc" directory.

There is still a lot of room for improvement regarding INSTALL_PREFIX.
See my notes in TODO.CMake.