Improved FindGTK3.cmake
There are various gtk3 builds/packages/bundles available on Windows. Some of them follow the
classic naming scheme (libs with -win32 suffix) and others prefer the original names.
Because we want to support as many packages as possible, we search for both naming styles.
Starting with the original names.

Tested with vcpkg (gtk+-3.22.19)
and Msys2 (gtk+-3.22.28)