Introduced support for Pthreads-win32 (a.k.a. pthreads4w)
Not my favorite port of Pthreads, but it's widely used.
I put a lot effort in assisting users with finding and linking with it.
See my comments in cmake/Modules/FindPTHREADS-WIN32.cmake

We will also be storing preprocessor definitions needed by
dependencies that link with libettercap in the variable EC_DEFINITIONS.