Tweaks for finding dn_expand()
We need dn_expand() so error if it's not found.

Windows has dn_expand built in mswsock. But we aren't using it.
We have our own ec_win_dn_expand() that does the job for us. So for now,
the success of the test is only meant for HAVE_DN_EXPAND to become True
(if it isn't already via HAVE_RESOLV).

http://www.sockets.com/mswsock.htm