Updated and extended PCRE support
Tons of improvements. FindPCRE.cmake can handle components and checks for PCRE_VERSION
the way it's supposed to.
On Windows PCRE is not only required but must also be at least at version 8.38.
We'll see maybe I set this lower. This is for the most part because we need
libettercap requires regcomp() and other regex functions which Windows doesn't
provide out-of-the-box.