Install app icon in XDG hicolor icon theme
Install the icon of the application in the hicolor XDG icon theme;
this way it can be properly loaded by XDG menus in the currently set
XDG icon theme, without looking in the legacy pixmaps location.

ICON_DIR is changed to refer to the top-level of the XDG hicolor icon
theme rather than the precise location for the application SVGs, so it
can be used in the future to install more icons in hicolor.