Use non-deprecated accessors for ASN1_STRING types
OpenSSL is going to make ASN1_STRING opaque soon, see:
https://github.com/openssl/openssl/issues/29117.

This changes your handling of the OCTET_STRING when setting
your extensions to use ASN1_STRING_get0_data, and to make a
copy of that data to change it.

As it appears you are setting HAVE_OPAQUE_RSA_DSA_DH for
OpenSSL versions > 1.1, The accessor in question is available
in all those versions you are targeting, and is also in BoringSSL
and LibreSSL. (Your legacy case here need not change for older
versions of OpenSSL).

As I was adding an allocation I converted the function to
single return.